tests/wpt/tests/css/css-fonts/parsing/font-synthesis-position-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing font-synthesis-position with valid values</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/7441">
<meta name="assert" content="font-synthesis-position supports the full grammar 'auto | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-synthesis-position', 'auto');
test_valid_value('font-synthesis-position', 'none');
</script>
</body>
</html>
